requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
black==25.1.0
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.10
django-debug-toolbar==5.0.1
django-extensions==3.2.3
django-htmx==1.22.0
django-widget-tweaks==1.5.0
flake8==7.1.2
idna==3.10
kombu==5.4.2
mccabe==0.7.0
mssql-django==1.5
mypy-extensions==1.0.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pycodestyle==2.12.1
pyflakes==3.2.0
pyodbc==5.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.3.0
vine==5.1.0
waitress==3.0.2
wcwidth==0.2.13
Whoosh==2.7.4
